requirements.txt
# Core dependencies
PyQt5>=5.15.0
pyqtgraph>=0.13.0
numpy>=1.21.0
scipy>=1.7.0

# CAN communication
python-can>=4.0.0
cantools>=39.0.0

# Data handling
pandas>=1.3.0

# Optional: Additional CAN interfaces
# python-can[pcan]  # For PCAN support on Windows
# python-can[ixxat]  # For IXXAT support

# Development dependencies
pytest>=7.0.0
pytest-qt>=4.0.0
black>=22.0.0
flake8>=4.0.0
